+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Remove firebase dependency
+* Allow log level to be applied
+* Allow HTTP client to be passed to requests
+
 ## 1.0.2
 
 * Improve pub score